backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
openai>=1.40.0
python-dotenv==1.0.0
google-generativeai==0.8.3
